+++ jodd-props/src/main/java/jodd/props/PropsParser.java
@@ -143,18 +143,26 @@ public class PropsParser implements Cloneable {
 
 			if (state == ParseState.COMMENT) {
 				// comment, skip to the end of the line
-				if (c == '\n') {
+				if (c == '\r') {
+					if ((ndx < len) && (in.charAt(ndx) == '\n')) {
+						ndx++;
+					}
+					state = ParseState.TEXT;
+				}
+				else if (c == '\n') {
 					state = ParseState.TEXT;
 				}
 			} else if (state == ParseState.ESCAPE) {
-				state = stateOnEscape;//ParseState.VALUE;
+				state = stateOnEscape;  //ParseState.VALUE;
 				switch (c) {
 					case '\r':
-					case '\n':
-						// if the EOL is \n or \r\n, escapes both chars
-						if (c == '\n') {
-							ndx--;
+						if ((ndx < len) && (in.charAt(ndx) == '\n')) {
+							ndx++;
 						}
+					case '\n':
+						// need to go 1 step back in order to escape
+						// the current line ending in the follow-up state
+						ndx--;
 						state = ParseState.ESCAPE_NEWLINE;
 						break;
 					// encode UTF character
@@ -277,8 +285,11 @@ public class PropsParser implements Cloneable {
 						break;
 
 					case '\r':
+						if ((ndx < len) && (in.charAt(ndx) == '\n')) {
+							ndx++;
+						}
 					case '\n':
-						if ((state == ParseState.ESCAPE_NEWLINE) && (c == '\n')) {
+						if (state == ParseState.ESCAPE_NEWLINE) {
 							sb.append(escapeNewLineValue);
 							if (!ignorePrefixWhitespacesOnNewLine) {
 								state = ParseState.VALUE;
@@ -296,7 +307,7 @@ public class PropsParser implements Cloneable {
 
 					case ' ':
 					case '\t':
-						if ((state == ParseState.ESCAPE_NEWLINE)) {
+						if (state == ParseState.ESCAPE_NEWLINE) {
 							break;
 						}
 					default: