+++ jodd-core/src/test/java/jodd/mutable/ValueTest.java
@@ -28,13 +28,20 @@ package jodd.mutable;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class ValueTest {
 
 	@Test
 	void testValue() {
 		Value<String> str = Value.of(null);
+
+		assertFalse(str.optional().isPresent());
+
 		str.set("123");
+
+		assertTrue(str.optional().isPresent());
 
 		assertEquals("123", str.get());